+++ playground/CHANGELOG.md
@@ -1,5 +1,26 @@
 # playground
 
+## 0.0.6
+
+### Patch Changes
+
+- Updated dependencies
+  - @gnosticdev/hono-actions@1.0.14
+
+## 0.0.5
+
+### Patch Changes
+
+- Updated dependencies
+  - @gnosticdev/hono-actions@1.0.13
+
+## 0.0.4
+
+### Patch Changes
+
+- Updated dependencies
+  - @gnosticdev/hono-actions@1.0.12
+
 ## 0.0.3
 
 ### Patch Changes